Refactor: Split UserContext into smaller files
Split UserContext.tsx into smaller files for better maintainability.
